Initial commit of snowy
Snowy is a document and content repository for saving blobs of binary
data in a consistent manor. Versioning of content can be achieved
through adding content to the same resource id.

Most of this work is a work in progress, so it is not done/finished yet,
so consider it as "use at your own risk!"