Query (#1)
* Introduction of querying documents

Currently broken

* Query integration tests

More tests around integration

* Store returns the correct order of entities

* Tags testing

This removes the flakey tests around tags and being in the wrong order
when they come out

* Introduction of the "coverage" story

This introduces the go test coverage from within the docker setup making
sure that it's repeatable. You can view the output in html form via:

    go tool cover -html=bin/coverage.out

* Improving test coverage

* Virtual store is now 100% covered

* Coverage testing

Improved real store running, making sure that the store is running
before testing starts

* Better example of mocking

Move the mock store into it's own folder to prevent issues with coverage.

* Api change for get, we can now pass a tag

* Generate mocks when building

Removal of the old mock files which turns out was a hacky way to
build them. At least this way, they will always be correct!

* Start testing of API service

This also includes changing documents interface to a struct so we can
deal with empty documents in a more coherant manor

* API tags testing

The following commit adds testing for tagging queries.

* Fix potential issues with the schema

The new schema is intended to represent the document schema correctly,
instead of a best guess.

* Get multiple integration tests

* Let's try and get travis working

* Make sure we have all the dependencies for mockgen

* Always return true

* We don't need to build the mocks everytime

* Make sure we don't clean the mocks when testing

* Now that travis is working, show the badge

* API documents insertion testing

* API coverage testing

* Removing the redirect

This commit adds more coverage of the code, but in reality it removes
the redirect if the url ends with a slash. It's not required and causes
confusion.

* Coverage tests for the API endpoints

* Adding testing of the documents themselves